// generate-data-script/src/shared/types.ts
export interface FlattenedProfile {
  // node-postgres serializes int8 to string as it's possibly bigger than Number.MAX_SAFE_NUMBER
  fid: number | string;
  address: string | null;
  username?: string | null;
  display_name?: string | null;
  avatar_url?: string | null;
  referrer?: string | null;
  avatar_verified?: boolean;
  followers?: number;
  following?: number;
  bio?: string | null;
  registered_at?: Date | string;
  updated_at?: Date | string;
}

export interface FlattenedCast {
  type: "text-short";
  published_at: Date;
  /** only tracks a full reindex, that is from discove/farcaster-indexer-v2/src/tasks/index-casts.ts **/
  last_reindexed_at: Date;
  recasted_cast_hash: string | null;
  is_recast: boolean;
  deleted: boolean;

  username: string | null;
  // node-postgres serializes int8 to string as it's possibly bigger than Number.MAX_SAFE_NUMBER
  author_fid: number | string;
  text: string;

  reply_parent_hash: string | null;
  reply_parent_username: string | null;
  // node-postgres serializes int8 to string as it's possibly bigger than Number.MAX_SAFE_NUMBER
  reply_parent_fid: number | null | string;

  hash: string;
  thread_hash: string;
  display_name: string | null;
  avatar_url: string | null;
  avatar_verified: boolean;
  mentions: JSON | any;

  replies: number | null;
  reactions: number | null;
  recasts: number | null;
  watches: number | null;

  recasters: JSON | any;
}

export type CustomCastMetrics = {
  new: number;
  top: number;
  casterengagement: number;
  toplikes: number;
  toprecasts: number;
  topwatches: number;
  topreactions: number;
  topreplies: number;
  topfollowers: number;
  topfollowernormalized: number;
  hot: number;
  hotfollowernormalized: number;
  hotfrequencynormalized: number;
};

// export interface EmailSubscription {
//   email: string;
//   subscriptions: {
//     weekly_cove_updates?: Record<string, boolean>;
//   };
//   is_suppressed: boolean;
//   created_at: string;
//   updated_at: string;
//   management_secret_token: string;
// }

/** this is a subset of the properties */
export interface OpenGraphType {
  url?: string;
  type?: string;
  image?: string;
  title?: string;
  locale?: string;
  favicon?: string;
  imageType?: string;
  site_name?: string;
  imageWidth?: string;
  description?: string;
  imageHeight?: string;
  articleModifiedTime?: string;
  video?: string;
  videoType?: string;
  videoWidth?: string;
  videoHeight?: string;
}

export interface FCCast extends FlattenedCast {
  reply_data?: FCCast[];
  og: {
    last_updated?: string;
    urls?: Array<{ url: string; last_status: number; tags: OpenGraphType }>;
  };
  reply_to_data?: FCCast;
  recast_data?: FCCast;
  weighted_keywords?: unknown;
  custom_metrics?: {
    custom_profile_metrics?: FCProfile["custom_metrics"];
    custom_cast_metrics?: CustomCastMetrics;
  };
}

export interface FCProfile extends FlattenedProfile {
  weighted_keywords?: unknown;
  // Datestring

  custom_metrics?: {
    // datestring
    last_known_active?: Date | string;
    total_casts: number;

    // registered_at
    new: number;
    total_recasts: number;
    recasts_score: number;
    // hard to track without tracking followers
    hot: number;
    hot_by_engagement: number;
    total_likes: number;
    // top?
    num_followers: number;
    num_following: number;
    frequency_score: number;
    engagement_score: number;
    per_follower_engagement_score: number;
  };
}