// generate-data-script/src/types/pg.d.ts
export interface ClientConfig {
  user?: string | undefined;
  database?: string | undefined;
  password?: string | (() => string | Promise<string>) | undefined;
  port?: number | undefined;
  host?: string | undefined;
  connectionString?: string | undefined;
  keepAlive?: boolean | undefined;
  stream?: any | undefined;
  statement_timeout?: false | number | undefined;
  parseInputDatesAsUTC?: boolean | undefined;
  ssl?: boolean | any | undefined;
  query_timeout?: number | undefined;
  keepAliveInitialDelayMillis?: number | undefined;
  idle_in_transaction_session_timeout?: number | undefined;
  application_name?: string | undefined;
  connectionTimeoutMillis?: number | undefined;
  types?: CustomTypesConfig | undefined;
  options?: string | undefined;
}

export type ConnectionConfig = ClientConfig;

export interface Defaults extends ClientConfig {
  poolSize?: number | undefined;
  poolIdleTimeout?: number | undefined;
  reapIntervalMillis?: number | undefined;
  binary?: boolean | undefined;
  parseInt8?: boolean | undefined;
}

export interface PoolConfig extends ClientConfig {
  // properties from module 'node-pool'
  max?: number | undefined;
  min?: number | undefined;
  idleTimeoutMillis?: number | undefined;
  log?: ((...messages: any[]) => void) | undefined;
  Promise?: PromiseConstructorLike | undefined;
  allowExitOnIdle?: boolean | undefined;
  maxUses?: number | undefined;
}

export interface QueryConfig<I extends any[] = any[]> {
  name?: string | undefined;
  text: string;
  values?: I | undefined;
  types?: CustomTypesConfig | undefined;
}

export interface CustomTypesConfig {
  getTypeParser: any;
}

export interface Submittable {
  submit: (connection: Connection) => void;
}

export interface QueryArrayConfig<I extends any[] = any[]>
  extends QueryConfig<I> {
  rowMode: "array";
}

export interface FieldDef {
  name: string;
  tableID: number;
  columnID: number;
  dataTypeID: number;
  dataTypeSize: number;
  dataTypeModifier: number;
  format: string;
}

export interface QueryResultBase {
  command: string;
  rowCount: number;
  oid: number;
  fields: FieldDef[];
}

export interface QueryResultRow {
  [column: string]: any;
}

export interface QueryResult<R extends QueryResultRow = any>
  extends QueryResultBase {
  rows: R[];
}

export interface QueryArrayResult<R extends any[] = any[]>
  extends QueryResultBase {
  rows: R[];
}

export interface Notification {
  processId: number;
  channel: string;
  payload?: string | undefined;
}
// ...
export class Connection {
  readonly stream: any;

  constructor(config?: ConnectionConfig);

  bind(config: BindConfig | null, more: boolean): void;
  execute(config: ExecuteConfig | null, more: boolean): void;
  parse(query: QueryParse, more: boolean): void;

  query(text: string): void;

  describe(msg: MessageConfig, more: boolean): void;
  close(msg: MessageConfig, more: boolean): void;

  flush(): void;
  sync(): void;
  end(): void;
}

/**
 * {@link https://node-postgres.com/api/pool}
 */
export class Pool {
  /**
   * Every field of the config object is entirely optional.
   * The config passed to the pool is also passed to every client
   * instance within the pool when the pool creates that client.
   */
  constructor(config?: PoolConfig);

  readonly totalCount: number;
  readonly idleCount: number;
  readonly waitingCount: number;

  connect(): Promise<PoolClient>;
  connect(
    callback: (
      err: Error,
      client: PoolClient,
      done: (release?: any) => void
    ) => void
  ): void;

  end(): Promise<void>;
  end(callback: () => void): void;

  query<T extends Submittable>(queryStream: T): T;
  // tslint:disable:no-unnecessary-generics
  query<R extends any[] = any[], I extends any[] = any[]>(
    queryConfig: QueryArrayConfig<I>,
    values?: I
  ): Promise<QueryArrayResult<R>>;
  query<R extends QueryResultRow = any, I extends any[] = any[]>(
    queryConfig: QueryConfig<I>
  ): Promise<QueryResult<R>>;
  query<R extends QueryResultRow = any, I extends any[] = any[]>(
    queryTextOrConfig: string | QueryConfig<I>,
    values?: I
  ): Promise<QueryResult<R>>;
  query<R extends any[] = any[], I extends any[] = any[]>(
    queryConfig: QueryArrayConfig<I>,
    callback: (err: Error, result: QueryArrayResult<R>) => void
  ): void;
  query<R extends QueryResultRow = any, I extends any[] = any[]>(
    queryTextOrConfig: string | QueryConfig<I>,
    callback: (err: Error, result: QueryResult<R>) => void
  ): void;
  query<R extends QueryResultRow = any, I extends any[] = any[]>(
    queryText: string,
    values: I,
    callback: (err: Error, result: QueryResult<R>) => void
  ): void;
  // tslint:enable:no-unnecessary-generics

  on(event: "error", listener: (err: Error, client: PoolClient) => void): this;
  on(
    event: "connect" | "acquire" | "remove",
    listener: (client: PoolClient) => void
  ): this;
}

export class ClientBase {
  constructor(config?: string | ClientConfig);

  connect(): Promise<void>;
  connect(callback: (err: Error) => void): void;

  query<T extends Submittable>(queryStream: T): T;
  // tslint:disable:no-unnecessary-generics
  query<R extends any[] = any[], I extends any[] = any[]>(
    queryConfig: QueryArrayConfig<I>,
    values?: I
  ): Promise<QueryArrayResult<R>>;
  query<R extends QueryResultRow = any, I extends any[] = any[]>(
    queryConfig: QueryConfig<I>
  ): Promise<QueryResult<R>>;
  query<R extends QueryResultRow = any, I extends any[] = any[]>(
    queryTextOrConfig: string | QueryConfig<I>,
    values?: I
  ): Promise<QueryResult<R>>;
  query<R extends any[] = any[], I extends any[] = any[]>(
    queryConfig: QueryArrayConfig<I>,
    callback: (err: Error, result: QueryArrayResult<R>) => void
  ): void;
  query<R extends QueryResultRow = any, I extends any[] = any[]>(
    queryTextOrConfig: string | QueryConfig<I>,
    callback: (err: Error, result: QueryResult<R>) => void
  ): void;
  query<R extends QueryResultRow = any, I extends any[] = any[]>(
    queryText: string,
    values: any[],
    callback: (err: Error, result: QueryResult<R>) => void
  ): void;
  // tslint:enable:no-unnecessary-generics

  copyFrom(queryText: string): any;
  copyTo(queryText: string): any;

  pauseDrain(): void;
  resumeDrain(): void;

  escapeIdentifier(str: string): string;
  escapeLiteral(str: string): string;

  on(event: "drain", listener: () => void): this;
  on(event: "error", listener: (err: Error) => void): this;
  on(event: "notice", listener: (notice: any) => void): this;
  on(event: "notification", listener: (message: Notification) => void): this;
  // tslint:disable-next-line unified-signatures
  on(event: "end", listener: () => void): this;
}

export class Client extends ClientBase {
  user?: string | undefined;
  database?: string | undefined;
  port: number;
  host: string;
  password?: string | undefined;
  ssl: boolean;

  constructor(config?: string | ClientConfig);

  end(): Promise<void>;
  end(callback: (err: Error) => void): void;
}

export interface PoolClient extends ClientBase {
  release(err?: Error | boolean): void;
}